broadcast sub-benchmarks
