+++ conn_test.go
@@ -13,7 +13,6 @@ import (
 	"io/ioutil"
 	"net"
 	"reflect"
-	"sync/atomic"
 	"testing"
 	"testing/iotest"
 	"time"
@@ -464,114 +463,3 @@ func TestFailedConnectionReadPanic(t *testing.T) {
 	}
 	t.Fatal("should not get here")
 }
-
-// broadcastBench allows to run broadcast benchmarks.
-// In every broadcast benchmark we create many connections, then broadcast message
-// to every connection and wait for all writes complete. This emulates an application
-// where many connections listen to the same data - i.e. PUB/SUB scenarios with many
-// subscribers in one channel.
-type broadcastBench struct {
-	w           io.Writer
-	message     *broadcastMessage
-	closeCh     chan struct{}
-	doneCh      chan struct{}
-	count       int32
-	conns       []*broadcastConn
-	compression bool
-	usePrepared bool
-}
-
-type broadcastMessage struct {
-	payload  []byte
-	prepared *PreparedMessage
-}
-
-type broadcastConn struct {
-	conn  *Conn
-	msgCh chan *broadcastMessage
-}
-
-func newBroadcastConn(c *Conn) *broadcastConn {
-	return &broadcastConn{
-		conn:  c,
-		msgCh: make(chan *broadcastMessage, 1),
-	}
-}
-
-func newBroadcastBench(usePrepared, compression bool) *broadcastBench {
-	bench := &broadcastBench{
-		w:           ioutil.Discard,
-		doneCh:      make(chan struct{}),
-		closeCh:     make(chan struct{}),
-		usePrepared: usePrepared,
-		compression: compression,
-	}
-	bm := &broadcastMessage{
-		payload: textMessages(1)[0],
-	}
-	if usePrepared {
-		pm, _ := NewPreparedMessage(TextMessage, bm.payload)
-		bm.prepared = pm
-	}
-	bench.message = bm
-	bench.makeConns(10000)
-	return bench
-}
-
-func (b *broadcastBench) makeConns(numConns int) {
-	conns := make([]*broadcastConn, numConns)
-
-	for i := 0; i < numConns; i++ {
-		c := newConn(fakeNetConn{Reader: nil, Writer: b.w}, true, 1024, 1024)
-		if b.compression {
-			c.enableWriteCompression = true
-			c.newCompressionWriter = compressNoContextTakeover
-		}
-		conns[i] = newBroadcastConn(c)
-		go func(c *broadcastConn) {
-			for {
-				select {
-				case msg := <-c.msgCh:
-					if b.usePrepared {
-						c.conn.WritePreparedMessage(msg.prepared)
-					} else {
-						c.conn.WriteMessage(TextMessage, msg.payload)
-					}
-					val := atomic.AddInt32(&b.count, 1)
-					if val%int32(numConns) == 0 {
-						b.doneCh <- struct{}{}
-					}
-				case <-b.closeCh:
-					return
-				}
-			}
-		}(conns[i])
-	}
-	b.conns = conns
-}
-
-func (b *broadcastBench) close() {
-	close(b.closeCh)
-}
-
-func (b *broadcastBench) runOnce() {
-	for _, c := range b.conns {
-		c.msgCh <- b.message
-	}
-	<-b.doneCh
-}
-
-func benchmarkBroadcast(b *testing.B, usePrepared, compression bool) {
-	bench := newBroadcastBench(usePrepared, compression)
-	defer bench.close()
-	b.ResetTimer()
-	for j := 0; j < b.N; j++ {
-		bench.runOnce()
-	}
-	b.ReportAllocs()
-}
-
-func BenchmarkBroadcastNoCompression(b *testing.B)           { benchmarkBroadcast(b, false, false) }
-func BenchmarkBroadcastWithCompression(b *testing.B)         { benchmarkBroadcast(b, false, true) }
-func BenchmarkBroadcastNoCompressionPrepared(b *testing.B)   { benchmarkBroadcast(b, true, false) }
-func BenchmarkBroadcastWithCompressionPrepared(b *testing.B) { benchmarkBroadcast(b, true, true) }
